Refactor : 중복되는 부분 공통 컴포넌트로 분리
